+++ database/migrations/2024_07_23_102018_create_reservations_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->foreignId('table_id')->constrained('tables');
             $table->dateTime('date');
-            $table->double('total', 10, 2);
+            $table->bigInteger('total');
             $table->timestamps();
         });
     }
